Generate aten_interned_strings.h automatically
This generates aten_interned_strings.h from `native_functions.yaml`
which is more like how it was originally done. The items deleted from
`interned_strings.h` are duplicates that need to be removed in order
for the code to compile, some of the remaining items may still be out
of date but it is fairly benign even if that's the case.

ghstack-source-id: 3c03615ba91218f08be86aa4e2058bdc94448dea
Pull Request resolved: https://github.com/pytorch/pytorch/pull/69407